feat(engine): failure reporting for external tasks
- in case a task cannot be completed by an external worker, a failure
  can be reported to the engine
- retries and retryTimeout can be specified externally when reporting the
  failure
- an incident is created when 0 retries are left (consistent with jobs)
- retries can be set via API to recover when retries are 0

related to CAM-4620

+++ engine/src/main/java/org/camunda/bpm/engine/impl/util/EnsureUtil.java
@@ -158,15 +158,15 @@ public final class EnsureUtil {
     }
   }
 
-  public static void ensureGreaterThanOrEqual(String variableName, int value1, int value2) {
+  public static void ensureGreaterThanOrEqual(String variableName, long value1, long value2) {
     ensureGreaterThanOrEqual("", variableName, value1, value2);
   }
 
-  public static void ensureGreaterThanOrEqual(String message, String variableName, int value1, int value2) {
+  public static void ensureGreaterThanOrEqual(String message, String variableName, long value1, long value2) {
     ensureGreaterThanOrEqual(ProcessEngineException.class, message, variableName, value1, value2);
   }
 
-  public static void ensureGreaterThanOrEqual(Class<? extends ProcessEngineException> exceptionClass, String message, String variableName, int value1, int value2) {
+  public static void ensureGreaterThanOrEqual(Class<? extends ProcessEngineException> exceptionClass, String message, String variableName, long value1, long value2) {
     if (value1 < value2) {
       throw generateException(exceptionClass, message, variableName, "is not greater than or equal to " + value2);
     }
